r45801 changed screenshots slightly, so adjust them
We adjust the reference screenshots for those cases that look like they
should pass.

Test crtc0 sometimes spuriously fails; the screenshot seems to be taken
when a rasterline in in the process of being drawn (if this is
possible?)

Test crtc10 seems to have regressed at some point in the past; at least
the reference screenshot looks more like the hardware screen photo than
the current code produces.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/testprogs@45806 379a1393-f5fb-40a0-bcee-ef074d9b53f7